# docs/swagger.yaml
basePath: /v1
definitions:
  controllers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.ClientDto:
    properties:
      birthDate:
        description: Date only
        format: date
        type: string
      email:
        description: E-mail string.
        example: mail@mail.com
        type: string
      info:
        description: 'Allowed characters: printable ASCII (Russian and English).'
        type: string
      name:
        description: Alphabetic characters only (Russian and English).
        maxLength: 255
        type: string
      patronymic:
        description: Alphabetic characters only (Russian and English).
        maxLength: 255
        type: string
      phone:
        description: E.164 phone number string.
        example: "+79876543210"
        type: string
      sex:
        description: Alphabetic characters only (Russian and English).
        type: string
      surname:
        description: Alphabetic characters only (Russian and English).
        maxLength: 255
        type: string
    type: object
  dto.DepartmentDto:
    properties:
      name:
        description: Alphabetic characters only (Russian and English).
        maxLength: 255
        type: string
      services:
        items:
          type: integer
        type: array
    required:
    - name
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Client:
    properties:
      birthDate:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      info:
        type: string
      name:
        type: string
      patronymic:
        type: string
      phone:
        type: string
      sex:
        type: string
      surname:
        type: string
      updated_at:
        type: string
    type: object
  models.Department:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      slug:
        type: string
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Service:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      categoryId:
        type: integer
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      departments:
        items:
          $ref: '#/definitions/models.Department'
        type: array
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.User:
    properties:
      active:
        type: boolean
      birthDate:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      departments:
        items:
          $ref: '#/definitions/models.Department'
        type: array
      email:
        type: string
      id:
        type: integer
      info:
        type: string
      name:
        type: string
      patronymic:
        type: string
      phone:
        type: string
      profession:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      roleId:
        type: integer
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      sex:
        type: string
      slug:
        type: string
      surname:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: vic.pozdeev@gmail.com
    name: Victor Pozdeev
  description: This is API specification for vet-clinic API server.
  license:
    name: MIT
    url: https://opensource.org/licenses/mit-license.php
  title: Vet clinic API
  version: v0.1.0
paths:
  /clients:
    get:
      consumes:
      - application/json
      description: Returns the list of clients.
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            items:
              $ref: '#/definitions/models.Client'
            type: array
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a client list.
      tags:
      - Clients
    post:
      consumes:
      - application/json
      description: Create a new client.
      parameters:
      - description: A new client data for creating.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ClientDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Create a new client. Required user''s role: Admin'
      tags:
      - Clients
  /clients/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the existing client.
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Delete the existing client. Required user''s role: Superuser'
      tags:
      - Clients
    get:
      consumes:
      - application/json
      description: Returns one record matched client's id.
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a client.
      tags:
      - Clients
    put:
      consumes:
      - application/json
      description: Update the existing client.
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: Client data for update.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ClientDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Update the existing client. Required user''s role: Admin'
      tags:
      - Clients
  /departments:
    get:
      consumes:
      - application/json
      description: Returns the list of departments.
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            items:
              $ref: '#/definitions/models.Department'
            type: array
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a department list.
      tags:
      - Departments
    post:
      consumes:
      - application/json
      description: Create a new department.
      parameters:
      - description: A new department data for creating.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.DepartmentDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Department'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Create a new department. Required user''s role: Owner'
      tags:
      - Departments
  /departments/{id_or_slug}:
    get:
      consumes:
      - application/json
      description: Returns one record matched department's id.
      parameters:
      - description: Department ID
        in: path
        name: id_or_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Department'
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a department.
      tags:
      - Departments
  /departments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the existing department.
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Department'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Delete the existing department. Required user''s role: Owner'
      tags:
      - Departments
    put:
      consumes:
      - application/json
      description: Update the existing department.
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: string
      - description: Department data for update.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.DepartmentDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Department'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Update the existing department. Required user''s role: Owner'
      tags:
      - Departments
...
schemes:
- http
swagger: "2.0"